more improvements to TOML Parm Parse
