+++ cmlm/utils/__init__.py
@@ -1,2 +1,2 @@
-from .input_file import TomlParmParse  # noqa: F401
+from .input_file import TomlParmParse # noqa: F401
 from .timer import Timer, TrackerBar  # noqa: F401
