+++ pyproject.toml
@@ -19,7 +19,7 @@ dependencies = [
     "pandas >= 2.0.3",
     "scikit-learn >= 1.7.0",
     "h5py >= 3.14.0",
-    "toml >= 0.10.0"
+    "tomlkit >= 0.13.0"
 ]
 
 [project.optional-dependencies]
